Use local storage only, never sync
Firefox requires an explicit addon ID in order to use sync storage in a
temporary addon, but the manifest.json "applications" key is invalid in
Chrome and will throw a warning. Avoiding this would require building
two versions of the extension, which would be too much work for the
absolutely tiny potential benefit it would bring. Local storage it is.